requirements.txt
python-telegram-bot==13.3
peewee==3.14.1
requests==2.26.0
